+++ src/main/java/Lab2/Calculator.java
@@ -9,15 +9,16 @@ package Lab2;
  * Все методы должны быть перегружены и работать с тремя
  * типами данных- double, long и int.
  */
+// todo cast
 public class Calculator {
 
   // Сумма
   private double sum(double x, double y){
-    return x + y;
+    return (double) (x + y);
   }
 
   private long sum(long x, long y){
-    return x + y;
+    return (long) (x + y);
   }
 
   private int sum(int x, int y){
@@ -27,12 +28,12 @@ public class Calculator {
   // Деление
   private double div(double x, double y) {
     checkZero((long) y);
-    return x / y;
+    return (double) (x / y);
   }
 
   private long div(long x, long y) {
     checkZero(y);
-    return x / y;
+    return (long) (x / y);
   }
 
   private int div(int x, int y) {
@@ -49,11 +50,11 @@ public class Calculator {
 
   // Умножение
   private double multipl(double x, double y) {
-    return x * y;
+    return (double) (x * y);
   }
 
   private long multipl(long x, long y) {
-    return x * y;
+    return (long) (x * y);
   }
 
   private int multipl(int x, int y) {
@@ -62,11 +63,11 @@ public class Calculator {
 
   // Вычитание
   private double subtraction(double x, double y){
-    return x - y;
+    return (double) (x - y);
   }
 
   private long subtraction(long x, long y){
-    return x - y;
+    return (long) (x - y);
   }
 
   private int subtraction(int x, int y){
